[scanner] fix: split feedback and stellar modal components
Fixes #21504

Fixes #21618

Fixes #21619

Fixes #21622

Fixes #21626
